Fixed sharps and flats "undefined" issue
Fixed an issue where sharps and flats would sometimes read as undefined as a result of not getting re-rounded to a 0-11 scale. If still not within the bounds, the note will once again add or subtract 12 to get to the proper value.

Also removed some old logging that was spamming the console.